app/requirements.txt
google-cloud==0.34.0
google-cloud-bigquery==3.12.0
google-cloud-aiplatform==1.35.0
tabulate==0.9.0
db-dtypes==1.1.1
python-dotenv==1.0.0
google-cloud-logging==3.8.0
streamlit==1.28.0
